// app/Models/Reg/Reports/StudCardData.ts
export class StudCardData{

    studId: number;
    parentId: number;
    studName: string;
    studLName: string;
    schoolId: number;
    schoolName: string;
    sectionId: number;
    sectionName: string;
    nationalityName: string;
    religionName: string;
    genderName: string;
    yearId: number;
    yearName: string;
    classId:number;
    className: string;
    classPrice: number;
    classSeqId: number;
    classSeqName: string;
    busNo: string;
    tourPrice: number;
    descountValue: number;
    birthDate:Date;
    birthPlace:string;
    genderId:number;
    image:string;
    email: string;
    approvedDate: Date;
    idNum: number;
    studStatusName: string;
    tel: string;
    fatherMobile: string;
    motherMobile: string;
    smsParentName: string;
    tourTypeIdName: string;
    busNote: string;
    studentBrotherValue: number;
    studentBrotherDescountType: string;
    totalValue: number;
    cityName: string;
    address: string;
    studMobile: string;
    parentWork: string;
    id:number;
    studCount:number;

    // joinYearId: number;
    // ApprovedId:number;
}