// app/Models/addLookups/schools/lkpSchool.ts
import { Lkplookup } from '../../Lookups/lkplookup';

export class LkpSchool {
      id: number;
      name: string ;
      cityId: number;
      citesLookup: Lkplookup;
      address: string;
      tel: string;
      mobile: string;
      fax: string;
      webPage: string;
      twitter: string;
      youTube: string;
      city: string;
      description: string;
      // engDescription: string;
      poBox: string;
      imageFile: any;
      imageId: number;
}